+++ merged-packages/smart-transactions-controller/setupJest.js
@@ -1,2 +1,0 @@
-// eslint-disable-next-line import-x/no-unassigned-import
-require('isomorphic-fetch');